Move a couple of functions to a common controller utils file
