+++ application-operator/controllers/metricstrait/metricstrait_utils_test.go
@@ -4,10 +4,11 @@
 package metricstrait
 
 import (
+	"testing"
+
 	"github.com/Jeffail/gabs/v2"
 	asserts "github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"testing"
 )
 
 // Test_updateStringMap tests metrics trait utility function updateStringMap
@@ -87,64 +88,6 @@ func Test_copyStringMapEntries(t *testing.T) {
 	assert.Len(output, 2)
 	assert.Len(source, 1)
 	assert.Len(target, 2)
-}
-
-// Test_stringSliceContainsString tests metrics trait utility function stringSliceContainsString
-func Test_stringSliceContainsString(t *testing.T) {
-	assert := asserts.New(t)
-	var slice []string
-	var find string
-	var found bool
-
-	// GIVEN a nil slice
-	// WHEN an empty string is searched for
-	// THEN verify false is returned
-	slice = nil
-	found = stringSliceContainsString(slice, find)
-	assert.Equal(found, false)
-
-	// GIVEN a slice with several strings
-	// WHEN one of the strings is searched for
-	// THEN verify string is found
-	slice = []string{"test-value-1", "test-value-2", "test-value-3"}
-	find = "test-value-2"
-	found = stringSliceContainsString(slice, find)
-	assert.Equal(found, true)
-
-	// GIVEN a slice with several strings
-	// WHEN a string not in the slice is searched for
-	// THEN verify string is not found
-	slice = []string{"test-value-1", "test-value-2", "test-value-3"}
-	find = "test-value-4"
-	found = stringSliceContainsString(slice, find)
-	assert.Equal(found, false)
-}
-
-// Test_removeStringFromStringSlice tests metrics trait utility function removeStringFromStringSlice
-func Test_removeStringFromStringSlice(t *testing.T) {
-	assert := asserts.New(t)
-	var slice []string
-	var remove string
-	var output []string
-
-	// GIVEN a nil slice and an empty string to remove
-	// WHEN the empty string is removed from the nil slice
-	// THEN verify that an empty slice is returned
-	slice = nil
-	remove = ""
-	output = removeStringFromStringSlice(slice, remove)
-	assert.NotNil(output)
-	assert.Len(output, 0)
-
-	// GIVEN a slice with several strings
-	// WHEN a string in the slice is removed
-	// THEN verify slice is correct
-	slice = []string{"test-value-1", "test-value-2", "test-value-3"}
-	remove = "test-value-2"
-	output = removeStringFromStringSlice(slice, remove)
-	assert.Equal("test-value-1", slice[0])
-	assert.Equal("test-value-2", slice[1])
-	assert.Len(output, 2)
 }
 
 // Test_getClusterNameFromObjectMetaOrDefault tests metrics trait utility function getClusterNameFromObjectMetaOrDefault
